- address sonar warnings - getPatientResourceTypes
